// app/generated/opentelemetry/proto/common/v1/common.ts
export interface AnyValue {
  stringValue?: string | undefined;
  boolValue?: boolean | undefined;
  intValue?: string | undefined;
  doubleValue?: number | undefined;
  arrayValue?: ArrayValue | undefined;
  kvlistValue?: KeyValueList | undefined;
  bytesValue?:
    | Uint8Array
    | undefined;
  /**
   * Reference to the string value in ProfilesDictionary.string_table.
   *
   * Note: This is currently used exclusively in the Profiling signal.
   * Implementers of OTLP receivers for signals other than Profiling should
   * treat the presence of this value as a non-fatal issue.
   * Log an error or warning indicating an unexpected field intended for the
   * Profiling signal and process the data as if this value were absent or
   * empty, ignoring its semantic content for the non-Profiling signal.
   *
   * Status: [Alpha]
   */
  stringValueStrindex?: number | undefined;
}

/**
 * ArrayValue is a list of AnyValue messages. We need ArrayValue as a message
 * since oneof in AnyValue does not allow repeated fields.
 */
export interface ArrayValue {
  /** Array of values. The array may be empty (contain 0 elements). */
  values?: AnyValue[] | undefined;
}

/**
 * KeyValueList is a list of KeyValue messages. We need KeyValueList as a message
 * since `oneof` in AnyValue does not allow repeated fields. Everywhere else where we need
 * a list of KeyValue messages (e.g. in Span) we use `repeated KeyValue` directly to
 * avoid unnecessary extra wrapping (which slows down the protocol). The 2 approaches
 * are semantically equivalent.
 */
export interface KeyValueList {
  /**
   * A collection of key/value pairs of key-value pairs. The list may be empty (may
   * contain 0 elements).
   *
   * The keys MUST be unique (it is not allowed to have more than one
   * value with the same key).
   * The behavior of software that receives duplicated keys can be unpredictable.
   */
  values?: KeyValue[] | undefined;
}

/**
 * Represents a key-value pair that is used to store Span attributes, Link
 * attributes, etc.
 */
export interface KeyValue {
  /**
   * The key name of the pair.
   * key_strindex MUST NOT be set if key is used.
   */
  key?:
    | string
    | undefined;
  /** The value of the pair. */
  value?:
    | AnyValue
    | undefined;
  /**
   * Reference to the string key in ProfilesDictionary.string_table.
   * key MUST NOT be set if key_strindex is used.
   *
   * Note: This is currently used exclusively in the Profiling signal.
   * Implementers of OTLP receivers for signals other than Profiling should
   * treat the presence of this key as a non-fatal issue.
   * Log an error or warning indicating an unexpected field intended for the
   * Profiling signal and process the data as if this value were absent or
   * empty, ignoring its semantic content for the non-Profiling signal.
   *
   * Status: [Alpha]
   */
  keyStrindex?: number | undefined;
}
// ...
export const AnyValue: MessageFns<AnyValue> = {
  fromJSON(object: any): AnyValue {
    return {
      stringValue: isSet(object.stringValue)
        ? globalThis.String(object.stringValue)
        : isSet(object.string_value)
        ? globalThis.String(object.string_value)
        : undefined,
      boolValue: isSet(object.boolValue)
        ? globalThis.Boolean(object.boolValue)
        : isSet(object.bool_value)
        ? globalThis.Boolean(object.bool_value)
        : undefined,
      intValue: isSet(object.intValue)
        ? globalThis.String(object.intValue)
        : isSet(object.int_value)
        ? globalThis.String(object.int_value)
        : undefined,
      doubleValue: isSet(object.doubleValue)
        ? globalThis.Number(object.doubleValue)
        : isSet(object.double_value)
        ? globalThis.Number(object.double_value)
        : undefined,
      arrayValue: isSet(object.arrayValue)
        ? ArrayValue.fromJSON(object.arrayValue)
        : isSet(object.array_value)
        ? ArrayValue.fromJSON(object.array_value)
        : undefined,
      kvlistValue: isSet(object.kvlistValue)
        ? KeyValueList.fromJSON(object.kvlistValue)
        : isSet(object.kvlist_value)
        ? KeyValueList.fromJSON(object.kvlist_value)
        : undefined,
      bytesValue: isSet(object.bytesValue)
        ? bytesFromBase64(object.bytesValue)
        : isSet(object.bytes_value)
        ? bytesFromBase64(object.bytes_value)
        : undefined,
      stringValueStrindex: isSet(object.stringValueStrindex)
        ? globalThis.Number(object.stringValueStrindex)
        : isSet(object.string_value_strindex)
        ? globalThis.Number(object.string_value_strindex)
        : undefined,
    };
  },

  toJSON(message: AnyValue): unknown {
    const obj: any = {};
    if (message.stringValue !== undefined) {
      obj.stringValue = message.stringValue;
    }
    if (message.boolValue !== undefined) {
      obj.boolValue = message.boolValue;
    }
    if (message.intValue !== undefined) {
      obj.intValue = message.intValue;
    }
    if (message.doubleValue !== undefined) {
      obj.doubleValue = message.doubleValue;
    }
    if (message.arrayValue !== undefined) {
      obj.arrayValue = ArrayValue.toJSON(message.arrayValue);
    }
    if (message.kvlistValue !== undefined) {
      obj.kvlistValue = KeyValueList.toJSON(message.kvlistValue);
    }
    if (message.bytesValue !== undefined) {
      obj.bytesValue = base64FromBytes(message.bytesValue);
    }
    if (message.stringValueStrindex !== undefined) {
      obj.stringValueStrindex = Math.round(message.stringValueStrindex);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AnyValue>, I>>(base?: I): AnyValue {
    return AnyValue.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AnyValue>, I>>(object: I): AnyValue {
    const message = createBaseAnyValue();
    message.stringValue = object.stringValue ?? undefined;
    message.boolValue = object.boolValue ?? undefined;
    message.intValue = object.intValue ?? undefined;
    message.doubleValue = object.doubleValue ?? undefined;
    message.arrayValue = (object.arrayValue !== undefined && object.arrayValue !== null)
      ? ArrayValue.fromPartial(object.arrayValue)
      : undefined;
    message.kvlistValue = (object.kvlistValue !== undefined && object.kvlistValue !== null)
      ? KeyValueList.fromPartial(object.kvlistValue)
      : undefined;
    message.bytesValue = object.bytesValue ?? undefined;
    message.stringValueStrindex = object.stringValueStrindex ?? undefined;
    return message;
  },
};

function createBaseArrayValue(): ArrayValue {
  return { values: [] };
}

export const ArrayValue: MessageFns<ArrayValue> = {
  fromJSON(object: any): ArrayValue {
    return {
      values: globalThis.Array.isArray(object?.values) ? object.values.map((e: any) => AnyValue.fromJSON(e)) : [],
    };
  },

  toJSON(message: ArrayValue): unknown {
    const obj: any = {};
    if (message.values?.length) {
      obj.values = message.values.map((e) => AnyValue.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ArrayValue>, I>>(base?: I): ArrayValue {
    return ArrayValue.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ArrayValue>, I>>(object: I): ArrayValue {
    const message = createBaseArrayValue();
    message.values = object.values?.map((e) => AnyValue.fromPartial(e)) || [];
    return message;
  },
};

function createBaseKeyValueList(): KeyValueList {
  return { values: [] };
}

export const KeyValueList: MessageFns<KeyValueList> = {
  fromJSON(object: any): KeyValueList {
    return {
      values: globalThis.Array.isArray(object?.values) ? object.values.map((e: any) => KeyValue.fromJSON(e)) : [],
    };
  },

  toJSON(message: KeyValueList): unknown {
    const obj: any = {};
    if (message.values?.length) {
      obj.values = message.values.map((e) => KeyValue.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<KeyValueList>, I>>(base?: I): KeyValueList {
    return KeyValueList.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<KeyValueList>, I>>(object: I): KeyValueList {
    const message = createBaseKeyValueList();
    message.values = object.values?.map((e) => KeyValue.fromPartial(e)) || [];
    return message;
  },
};

function createBaseKeyValue(): KeyValue {
  return { key: "", value: undefined, keyStrindex: 0 };
}

export const KeyValue: MessageFns<KeyValue> = {
  fromJSON(object: any): KeyValue {
    return {
      key: isSet(object.key) ? globalThis.String(object.key) : "",
      value: isSet(object.value) ? AnyValue.fromJSON(object.value) : undefined,
      keyStrindex: isSet(object.keyStrindex)
        ? globalThis.Number(object.keyStrindex)
        : isSet(object.key_strindex)
        ? globalThis.Number(object.key_strindex)
        : 0,
    };
  },

  toJSON(message: KeyValue): unknown {
    const obj: any = {};
    if (message.key !== undefined && message.key !== "") {
      obj.key = message.key;
    }
    if (message.value !== undefined) {
      obj.value = AnyValue.toJSON(message.value);
    }
    if (message.keyStrindex !== undefined && message.keyStrindex !== 0) {
      obj.keyStrindex = Math.round(message.keyStrindex);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<KeyValue>, I>>(base?: I): KeyValue {
    return KeyValue.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<KeyValue>, I>>(object: I): KeyValue {
    const message = createBaseKeyValue();
    message.key = object.key ?? "";
    message.value = (object.value !== undefined && object.value !== null)
      ? AnyValue.fromPartial(object.value)
      : undefined;
    message.keyStrindex = object.keyStrindex ?? 0;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}